Leaving out distutils.utils. Breaks sanic.
New version of Sanic uses distutils.utils.strtobool.
Hence I'm going to leave distutils.utils in place for now
and file a PR to Sanic to reimlement those 3 lines of code.
I understand it's part of stdlib and it's natual to borrow
it but it looks odd to use distutils in mainline code.